% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_category_coverage.R
\name{calc_category_coverage}
\alias{calc_category_coverage}
\title{Calculate Genomic Coverage of Annotation Categories}
\usage{
calc_category_coverage(bed, category_list = NULL, bp = 2489385779)
}
\arguments{
\item{bed}{A \code{data.frame} or \code{data.table} containing functional annotations from a BED file.
The first four columns are chr, start, end, and category.}

\item{category_list}{An optional character vector specifying the priority order of categories.
Genomic regions will be assigned to mutually exclusive categories, where categories listed earlier take precedence.
A region in multiple categories will be assigned to the category with the highest priority.}

\item{bp}{A numeric value specifying the total genome length in base pairs (default is 2489385779, for cattle autosomes).}
}
\value{
A \code{data.table} containing the proportion of genomic sequence covered by each functional annotation category.
}
\description{
This function calculates the proportion of genomic sequence covered by each functional annotation category, relative to the total genome length.
}
\examples{
bed <- data.frame(
  chr = c("1", "1", "2", "1", "2"),
  start = c(1000, 1500, 2000, 1800, 2200),
  end = c(2000, 3000, 2500, 2200, 2500),
  category = c("CDS", "intron", "CDS", "CDS", "intron")
)
# Calculate the proportion of genome covered by each category
cat_prop = calc_category_coverage(bed, category_list = c("CDS", "intron"))

}
