% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dairy_example}
\alias{dairy_example}
\title{Example dairy cattle data}
\format{
A list containing four data tables:
\describe{
\item{bfmap}{Genetic fine-mapping summary data generated by BFMAP (forward selection)}
\item{snp2annot}{SNP functional annotations with CDS, UTR, promoter, and intron categories}
\item{cat_prop}{Proportions of whole-genome SNPs in functional annotation categories}
\item{gene_annot}{Gene annotations with genomic positions}
}
}
\source{
TBA (2024+)
}
\usage{
dairy_example
}
\description{
Sample dataset from our dairy cattle genetic fine-mapping study (2024+).
}
\keyword{datasets}
