% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_snp_annotation.R
\name{map_snp_annotation}
\alias{map_snp_annotation}
\title{Map SNPs to Functional Annotations}
\usage{
map_snp_annotation(bfmap, bed, category_list = NULL, multi_cat = FALSE)
}
\arguments{
\item{bfmap}{A \code{data.frame} or \code{data.table} containing fine-mapping summary statistics generated by BFMAP (forward selection).
Required columns are Chr, Pos, and SNPname.}

\item{bed}{A \code{data.frame} or \code{data.table} containing functional annotations from a BED file.
The first four columns are chr, start, end, and category.}

\item{category_list}{An optional character vector specifying the priority order of categories for SNP assignment.
Categories listed earlier will take precedence.
A SNP in multiple categories will be assigned to the category with the highest priority.}

\item{multi_cat}{A logical value indicating whether to assign SNPs to each category independently or to a set of mutually exclusive categories.
If FALSE (default), SNPs will be assigned to each category independently, resulting in binary indicators for each category.
If TRUE, SNPs will be assigned to a set of mutually exclusive categories, and category_list is required to provide the priority order.}
}
\value{
A \code{data.table} containing SNP-to-annotation mappings. For multi_cat = FALSE,
the output includes a binary indicator column for each category. For multi_cat = TRUE,
it includes a single category assignment column based on the provided priority order.
}
\description{
This function maps SNPs to functional annotations based on their genomic positions.
It takes fine-mapping summary data and functional annotation BED data as input and
outputs a data table of SNP-to-annotation mappings.
}
\examples{
bfmap <- data.frame(
  Chr = c("1", "1", "2"),
  Pos = c(1000, 2000, 1500),
  SNPname = c("rs1", "rs2", "rs3")
)
bed <- data.frame(
  chr = c("1", "1", "2"),
  start = c(500, 1800, 1400),
  end = c(2200, 2200, 1600),
  category = c("CDS", "intron", "CDS")
)
snp2annot = map_snp_annotation(bfmap, bed)
snp2annot = map_snp_annotation(bfmap, bed, category_list = c("CDS", "intron"), multi_cat = TRUE)

}
